GenerateOrderV2 Code refactor (#6342)
