+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/orders/src/configs/generateOrdersConstants.js
@@ -1,0 +1,315 @@
+import {
+  configCheckout,
+  configRejectSubmission,
+  configsAssignDateToRescheduledHearing,
+  configsAssignNewHearingDate,
+  configsBail,
+  configsCaseTransferAccept,
+  configsCaseTransferReject,
+  configsCaseWithdrawalAccept,
+  configsCaseWithdrawalReject,
+  configsCreateOrderWarrant,
+  configsInitiateRescheduleHearingDate,
+  configsIssueNotice,
+  configsIssueSummons,
+  configsJudgement,
+  configsOrderMandatorySubmissions,
+  configsOrderSection202CRPC,
+  configsOrderSubmissionExtension,
+  configsOrderTranferToADR,
+  configsOthers,
+  configsRejectCheckout,
+  configsRejectRescheduleHeadingDate,
+  configsRescheduleHearingDate,
+  configsScheduleHearingDate,
+  configsScheduleNextHearingDate,
+  configsVoluntarySubmissionStatus,
+  configsIssueBailAcceptance,
+  configsIssueBailReject,
+  configsSetTermBail,
+  configsAcceptRejectDelayCondonation,
+  configsAdmitCase,
+  configsDismissCase,
+  configsApproveRejectLitigantDetailsChange,
+  replaceAdvocateConfig,
+  configsCreateOrderProclamation,
+  configsCreateOrderAttachment,
+  configsMoveCaseToLongPendingRegister,
+  configsMoveCaseOutOfLongPendingRegister,
+  configsCost,
+  configsWitnessBatta,
+  configsCaseSettlementAccept,
+  configsCaseSettlementReject,
+  configsAbateCase,
+  configAcceptReschedulingRequest,
+  configMiscellaneousProcess,
+} from "./ordersCreateConfig";
+
+export const configKeys = {
+  SECTION_202_CRPC: configsOrderSection202CRPC,
+  MANDATORY_SUBMISSIONS_RESPONSES: configsOrderMandatorySubmissions,
+  EXTENSION_OF_DOCUMENT_SUBMISSION_DATE: configsOrderSubmissionExtension,
+  REFERRAL_CASE_TO_ADR: configsOrderTranferToADR,
+  SCHEDULE_OF_HEARING_DATE: configsScheduleHearingDate,
+  SCHEDULING_NEXT_HEARING: configsScheduleNextHearingDate,
+  RESCHEDULE_OF_HEARING_DATE: configsRescheduleHearingDate,
+  CHECKOUT_ACCEPTANCE: configCheckout,
+  CHECKOUT_REJECT: configsRejectCheckout,
+  REJECTION_RESCHEDULE_REQUEST: configsRejectRescheduleHeadingDate,
+  INITIATING_RESCHEDULING_OF_HEARING_DATE: configsInitiateRescheduleHearingDate,
+  ASSIGNING_DATE_RESCHEDULED_HEARING: configsAssignDateToRescheduledHearing,
+  ASSIGNING_NEW_HEARING_DATE: configsAssignNewHearingDate,
+  CASE_TRANSFER_ACCEPT: configsCaseTransferAccept,
+  CASE_TRANSFER_REJECT: configsCaseTransferReject,
+  SETTLEMENT_ACCEPT: configsCaseSettlementAccept,
+  SETTLEMENT_REJECT: configsCaseSettlementReject,
+  SUMMONS: configsIssueSummons,
+  NOTICE: configsIssueNotice,
+  BAIL: configsBail,
+  WARRANT: configsCreateOrderWarrant,
+  PROCLAMATION: configsCreateOrderProclamation,
+  ATTACHMENT: configsCreateOrderAttachment,
+  WITHDRAWAL_ACCEPT: configsCaseWithdrawalAccept,
+  WITHDRAWAL_REJECT: configsCaseWithdrawalReject,
+  OTHERS: configsOthers,
+  APPROVE_VOLUNTARY_SUBMISSIONS: configsVoluntarySubmissionStatus,
+  REJECT_VOLUNTARY_SUBMISSIONS: configRejectSubmission,
+  JUDGEMENT: configsJudgement,
+  REJECT_BAIL: configsIssueBailReject,
+  ACCEPT_BAIL: configsIssueBailAcceptance,
+  SET_BAIL_TERMS: configsSetTermBail,
+  ACCEPTANCE_REJECTION_DCA: configsAcceptRejectDelayCondonation,
+  TAKE_COGNIZANCE: configsAdmitCase,
+  DISMISS_CASE: configsDismissCase,
+  APPROVAL_REJECTION_LITIGANT_DETAILS_CHANGE: configsApproveRejectLitigantDetailsChange,
+  ADVOCATE_REPLACEMENT_APPROVAL: replaceAdvocateConfig,
+  MOVE_CASE_TO_LONG_PENDING_REGISTER: configsMoveCaseToLongPendingRegister,
+  MOVE_CASE_OUT_OF_LONG_PENDING_REGISTER: configsMoveCaseOutOfLongPendingRegister,
+  COST: configsCost,
+  WITNESS_BATTA: configsWitnessBatta,
+  ABATE_CASE: configsAbateCase,
+  ACCEPT_RESCHEDULING_REQUEST: configAcceptReschedulingRequest,
+  MISCELLANEOUS_PROCESS: configMiscellaneousProcess,
+};
+
+export const stateSlaMap = {
+  SECTION_202_CRPC: 3,
+  MANDATORY_SUBMISSIONS_RESPONSES: 3,
+  EXTENSION_OF_DOCUMENT_SUBMISSION_DATE: 3,
+  REFERRAL_CASE_TO_ADR: 3,
+  SCHEDULE_OF_HEARING_DATE: 3,
+  RESCHEDULE_OF_HEARING_DATE: 3,
+  REJECTION_RESCHEDULE_REQUEST: 3,
+  APPROVAL_RESCHEDULE_REQUEST: 3,
+  INITIATING_RESCHEDULING_OF_HEARING_DATE: 1,
+  ASSIGNING_DATE_RESCHEDULED_HEARING: 3,
+  ASSIGNING_NEW_HEARING_DATE: 3,
+  CASE_TRANSFER_ACCEPT: 3,
+  CASE_TRANSFER_REJECT: 3,
+  SETTLEMENT_ACCEPT: 3,
+  SETTLEMENT_REJECT: 3,
+  SUMMONS: 3,
+  NOTICE: 3,
+  BAIL: 3,
+  WARRANT: 3,
+  PROCLAMATION: 3,
+  ATTACHMENT: 3,
+  WITHDRAWAL_ACCEPT: 3,
+  WITHDRAWAL_REJECT: 3,
+  OTHERS: 3,
+  APPROVE_VOLUNTARY_SUBMISSIONS: 3,
+  REJECT_VOLUNTARY_SUBMISSIONS: 3,
+  REJECT_BAIL: 3,
+  ACCEPT_BAIL: 3,
+  SET_BAIL_TERMS: 3,
+  JUDGEMENT: 3,
+  CHECKOUT_ACCEPTANCE: 1,
+  CHECKOUT_REJECT: 1,
+  COST: 3,
+  WITNESS_BATTA: 3,
+  DRAFT_IN_PROGRESS: 2,
+  ABATE_CASE: 3,
+  ACCEPT_RESCHEDULING_REQUEST: 3,
+  MISCELLANEOUS_PROCESS: 3,
+};
+
+export const dayInMillisecond = 24 * 3600 * 1000;
+export const ErrorAttendeesKey = "attendees";
+
+/**
+ * ORDER_TYPE_SETS
+ * Base order type codes for each case scenario.
+ * "SCHEDULE_OF_HEARING_DATE" is NOT included here — the resolver adds it
+ * automatically when there is no in-progress / active hearing.
+ */
+export const ORDER_TYPE_SETS = {
+  // PENDING_RESPONSE / PENDING_ADMISSION + delay application pending
+  PENDING_DELAY: [
+    "DISMISS_CASE",
+    "SUMMONS",
+    "NOTICE",
+    "SECTION_202_CRPC",
+    "MANDATORY_SUBMISSIONS_RESPONSES",
+    "REFERRAL_CASE_TO_ADR",
+    "WARRANT",
+    "OTHERS",
+    "JUDGEMENT",
+    "ACCEPT_BAIL",
+    "PROCLAMATION",
+    "ATTACHMENT",
+    "COST",
+    "WITNESS_BATTA",
+    "ABATE_CASE",
+    "MISCELLANEOUS_PROCESS",
+  ],
+  // PENDING_RESPONSE / PENDING_ADMISSION + bail application pending
+  PENDING_BAIL: [
+    "TAKE_COGNIZANCE",
+    "DISMISS_CASE",
+    "SUMMONS",
+    "NOTICE",
+    "SECTION_202_CRPC",
+    "MANDATORY_SUBMISSIONS_RESPONSES",
+    "REFERRAL_CASE_TO_ADR",
+    "WARRANT",
+    "OTHERS",
+    "JUDGEMENT",
+    "PROCLAMATION",
+    "ATTACHMENT",
+    "COST",
+    "WITNESS_BATTA",
+    "ABATE_CASE",
+    "MISCELLANEOUS_PROCESS",
+  ],
+  // PENDING_RESPONSE / PENDING_ADMISSION — default
+  PENDING_DEFAULT: [
+    "TAKE_COGNIZANCE",
+    "DISMISS_CASE",
+    "SUMMONS",
+    "NOTICE",
+    "SECTION_202_CRPC",
+    "MANDATORY_SUBMISSIONS_RESPONSES",
+    "REFERRAL_CASE_TO_ADR",
+    "WARRANT",
+    "OTHERS",
+    "JUDGEMENT",
+    "ACCEPT_BAIL",
+    "PROCLAMATION",
+    "ATTACHMENT",
+    "COST",
+    "WITNESS_BATTA",
+    "ABATE_CASE",
+    "MISCELLANEOUS_PROCESS",
+  ],
+  // courtCaseNumber + isLPRCase
+  ADMITTED_LPR: [
+    "SUMMONS",
+    "NOTICE",
+    "MANDATORY_SUBMISSIONS_RESPONSES",
+    "WARRANT",
+    "OTHERS",
+    "ACCEPT_BAIL",
+    "PROCLAMATION",
+    "ATTACHMENT",
+    "MOVE_CASE_OUT_OF_LONG_PENDING_REGISTER",
+    "COST",
+    "WITNESS_BATTA",
+    "MISCELLANEOUS_PROCESS",
+  ],
+  // courtCaseNumber + !lprNumber (standard admitted, eligible for LPR)
+  ADMITTED_NO_LPR: [
+    "SUMMONS",
+    "NOTICE",
+    "DISMISS_CASE",
+    "SECTION_202_CRPC",
+    "MANDATORY_SUBMISSIONS_RESPONSES",
+    "REFERRAL_CASE_TO_ADR",
+    "WARRANT",
+    "OTHERS",
+    "JUDGEMENT",
+    "ACCEPT_BAIL",
+    "PROCLAMATION",
+    "ATTACHMENT",
+    "MOVE_CASE_TO_LONG_PENDING_REGISTER",
+    "COST",
+    "WITNESS_BATTA",
+    "ABATE_CASE",
+    "MISCELLANEOUS_PROCESS",
+  ],
+  // courtCaseNumber — already has lprNumber (general admitted)
+  ADMITTED_DEFAULT: [
+    "SUMMONS",
+    "NOTICE",
+    "DISMISS_CASE",
+    "SECTION_202_CRPC",
+    "MANDATORY_SUBMISSIONS_RESPONSES",
+    "REFERRAL_CASE_TO_ADR",
+    "WARRANT",
+    "OTHERS",
+    "JUDGEMENT",
+    "ACCEPT_BAIL",
+    "PROCLAMATION",
+    "ATTACHMENT",
+    "COST",
+    "WITNESS_BATTA",
+    "ABATE_CASE",
+    "MISCELLANEOUS_PROCESS",
+  ],
+  // Fallback (no matching case status)
+  FALLBACK: [
+    "SUMMONS",
+    "NOTICE",
+    "DISMISS_CASE",
+    "SECTION_202_CRPC",
+    "MANDATORY_SUBMISSIONS_RESPONSES",
+    "REFERRAL_CASE_TO_ADR",
+    "SCHEDULE_OF_HEARING_DATE", // always show schedule in fallback
+    "WARRANT",
+    "OTHERS",
+    "JUDGEMENT",
+    "ACCEPT_BAIL",
+    "PROCLAMATION",
+    "ATTACHMENT",
+    "COST",
+    "WITNESS_BATTA",
+    "ABATE_CASE",
+    "MISCELLANEOUS_PROCESS",
+  ],
+};
+
+/**
+ * Creates a default order data object with initial values.
+ * This is used when creating a new order from scratch.
+ *
+ * @param {Object} params - Parameters for creating default order data
+ * @param {string} params.tenantId - Tenant ID
+ * @param {string} params.cnrNumber - CNR number of the case
+ * @param {string} params.filingNumber - Filing number of the case
+ * @returns {Object} Default order data object
+ */
+export const createDefaultOrderData = ({ tenantId, cnrNumber, filingNumber }) => {
+  // Import OrderWorkflowAction inline to avoid circular dependencies
+  const OrderWorkflowAction = { SAVE_DRAFT: "SAVE_DRAFT" };
+  
+  return {
+    createdDate: null,
+    tenantId,
+    cnrNumber,
+    filingNumber,
+    orderCategory: "INTERMEDIATE",
+    statuteSection: {
+      tenantId,
+    },
+    status: "",
+    isActive: true,
+    workflow: {
+      action: OrderWorkflowAction.SAVE_DRAFT,
+      comments: "Creating order",
+      assignes: [],
+      rating: null,
+      documents: [{}],
+    },
+    documents: [],
+    additionalDetails: { formdata: {} },
+  };
+};
