+++ frontend/micro-ui/web/micro-ui-internals/packages/modules/orders/src/utils/orderUtils.js
@@ -1,7 +1,8 @@
 import { CloseSvg } from "@egovernments/digit-ui-components";
 import React from "react";
-import { formatAddress, mapAddressDetails } from ".";
+import { formatAddress, mapAddressDetails, getComplainantName, getRespondantName } from ".";
 import { DateUtils } from "@egovernments/digit-ui-module-dristi/src/Utils";
+import { getCourtFee } from "./orderApiCallUtils";
 
 export const Heading = (props) => {
   return <h1 className="heading-m">{props.label}</h1>;
@@ -77,10 +78,10 @@ export const getUpdateDocuments = (documents, documentsFile, signedDoucumentUplo
       return documents?.map((doc) =>
         doc?.documentType === "UNSIGNED"
           ? {
-              ...doc,
-              fileStore: documentsFile?.fileStore,
-              additionalDetails: documentsFile?.additionalDetails,
-            }
+            ...doc,
+            fileStore: documentsFile?.fileStore,
+            additionalDetails: documentsFile?.additionalDetails,
+          }
           : doc
       );
     }
@@ -134,9 +135,8 @@ export const generateAddress = ({
   if (address) {
     return address;
   }
-  return `${locality ? `${locality},` : ""} ${district ? `${district},` : ""} ${city ? `${city},` : ""} ${state ? `${state},` : ""} ${
-    pincode ? `- ${pincode}` : ""
-  }`.trim();
+  return `${locality ? `${locality},` : ""} ${district ? `${district},` : ""} ${city ? `${city},` : ""} ${state ? `${state},` : ""} ${pincode ? `- ${pincode}` : ""
+    }`.trim();
 };
 
 export const channelTypeEnum = {
@@ -191,8 +191,8 @@ export const getParties = (type, orderSchema, allParties) => {
     parties = orderSchema?.orderDetails?.respondentName?.name
       ? [orderSchema?.orderDetails?.respondentName?.name]
       : orderSchema?.orderDetails?.respondentName
-      ? [orderSchema?.orderDetails?.respondentName]
-      : [];
+        ? [orderSchema?.orderDetails?.respondentName]
+        : [];
   } else if (["SUMMONS", "NOTICE"].includes(type)) {
     parties = orderSchema?.orderDetails?.respondentName;
   } else if (type === "SECTION_202_CRPC") {
@@ -782,4 +782,402 @@ export const _getTaskPayload = (taskCaseDetails, orderData, filingDate, schedule
   });
 
   return payload;
+};
+
+export const getRaiseBailBondReferenceId = ({ accusedKey, filingNumber }) => {
+  try {
+    const safeAccused = `_ACC_${accusedKey || "UNKNOWN"}`;
+    return `MANUAL_RAISE_BAIL_BOND_${filingNumber}${safeAccused}`;
+  } catch (e) {
+    console.error(e);
+    return `MANUAL_RAISE_BAIL_BOND_${filingNumber}_ACC_UNKNOWN`;
+  }
+};
+
+export const createTaskPayload = async (
+  orderType,
+  orderDetails,
+  { caseDetails, courtRoomData, tenantId, judgeName }
+) => {
+  let payload = {};
+  const { litigants } = caseDetails;
+  const complainantIndividualId = litigants?.find((item) => item?.partyType === "complainant.primary")?.individualId;
+
+  const orderData = orderDetails?.order;
+  const orderFormData = getFormData(orderType, orderData);
+  const orderFormValue = orderDetails?.order?.additionalDetails?.formdata;
+  const respondentNameData = getOrderData(orderType, orderFormData);
+  const formDataKeyMap = {
+    NOTICE: "noticeOrder",
+    SUMMONS: "SummonsOrder",
+    WARRANT: "warrantFor",
+    PROCLAMATION: "proclamationFor",
+    ATTACHMENT: "attachmentFor",
+  };
+  const selectedChannel = orderData?.additionalDetails?.formdata?.[formDataKeyMap[orderType]]?.selectedChannels;
+  const noticeType = orderData?.additionalDetails?.formdata?.noticeType?.type;
+  const respondentAddress = orderFormData?.addressDetails
+    ? orderFormData?.addressDetails?.map((data) => ({ ...data?.addressDetails }))
+    : respondentNameData?.address
+      ? respondentNameData?.address
+      : caseDetails?.additionalDetails?.respondentDetails?.formdata?.[0]?.data?.addressDetails?.map((data) => data?.addressDetails);
+  const partyIndex = orderFormData?.party?.data?.partyIndex || "";
+  const result = getRespondantName(respondentNameData);
+  const respondentName = result?.name || result;
+  const respondentPhoneNo = orderFormData?.party?.data?.phone_numbers || [];
+  const respondentEmail = orderFormData?.party?.data?.email || [];
+  const respondentUniqueId = orderFormData?.party?.data?.uniqueId || orderFormData?.party?.uniqueId || "";
+  const complainantDetails = caseDetails?.additionalDetails?.complainantDetails?.formdata?.find(
+    (d) => d?.data?.complainantVerification?.individualDetails?.individualId === complainantIndividualId
+  )?.data;
+
+  const state = complainantDetails?.addressDetails?.state || "";
+  const district = complainantDetails?.addressDetails?.district || "";
+  const city = complainantDetails?.addressDetails?.city || "";
+  const pincode = complainantDetails?.addressDetails?.pincode || "";
+  const latitude = complainantDetails?.addressDetails?.pincode?.latitude || "";
+  const longitude = complainantDetails?.addressDetails?.pincode?.longitude || "";
+  const complainantName = getComplainantName(complainantDetails);
+  const locality = complainantDetails?.addressDetails?.locality || "";
+  const complainantAddress = {
+    pincode: pincode,
+    district: district,
+    city: city,
+    state: state,
+    coordinate: {
+      longitude: longitude,
+      latitude: latitude,
+    },
+    locality: locality,
+  };
+  const courtDetails = courtRoomData?.Court_Rooms?.find((data) => data?.code === caseDetails?.courtId);
+  const ownerType = orderFormData?.party?.data?.ownerType;
+
+  const respondentDetails = {
+    name: respondentName,
+    address: { ...respondentAddress?.[0], coordinate: respondentAddress?.[0]?.coordinates },
+    phone: respondentPhoneNo[0] || "",
+    email: respondentEmail[0] || "",
+    age: "",
+    gender: "",
+    uniqueId: respondentUniqueId,
+    ...(ownerType && { ownerType: ownerType }),
+  };
+  const caseRespondent = {
+    name: caseDetails?.additionalDetails?.respondentDetails?.formdata?.[0]?.data?.respondentFirstName || "",
+    address: caseDetails?.additionalDetails?.respondentDetails?.formdata?.[0]?.data?.addressDetails?.[0]?.addressDetails,
+    phone: caseDetails?.additionalDetails?.respondentDetails?.formdata?.[0]?.data?.phonenumbers?.mobileNumber?.[0] || "",
+    email: caseDetails?.additionalDetails?.respondentDetails?.formdata?.[0]?.data?.emails?.emailId?.[0] || "",
+    age: caseDetails?.additionalDetails?.respondentDetails?.formdata?.[0]?.data?.respondentAge,
+    gender: "",
+  };
+
+  switch (orderType) {
+    case "SUMMONS":
+      payload = {
+        summonDetails: {
+          issueDate: orderData?.auditDetails?.lastModifiedTime,
+          caseFilingDate: caseDetails?.filingDate,
+          docSubType: orderFormData?.party?.data?.partyType === "Witness" ? "WITNESS" : "ACCUSED",
+        },
+        respondentDetails: orderFormData?.party?.data?.partyType === "Witness" ? caseRespondent : respondentDetails,
+        ...(orderFormData?.party?.data?.partyType === "Witness" && { witnessDetails: respondentDetails }),
+        complainantDetails: {
+          name: complainantName,
+          address: complainantAddress,
+        },
+        caseDetails: {
+          caseTitle: caseDetails?.caseTitle,
+          year: new Date(caseDetails).getFullYear(),
+          hearingDate: new Date(orderData?.additionalDetails?.formdata?.dateForHearing || "").getTime(),
+          courtName: courtDetails?.name,
+          courtAddress: courtDetails?.address,
+          courtPhone: courtDetails?.phone,
+          courtId: caseDetails?.courtId,
+          hearingNumber: orderData?.hearingNumber,
+          judgeName: judgeName,
+        },
+        deliveryChannels: {
+          channelName: "",
+          status: "",
+          statusChangeDate: "",
+          fees: 0,
+          feesStatus: "pending",
+        },
+      };
+      break;
+    case "NOTICE":
+      payload = {
+        noticeDetails: {
+          issueDate: orderData?.auditDetails?.lastModifiedTime,
+          caseFilingDate: caseDetails?.filingDate,
+          noticeType,
+          docSubType: orderFormData?.party?.data?.partyType === "Witness" ? "WITNESS" : "ACCUSED",
+          partyIndex: partyIndex,
+        },
+        respondentDetails: orderFormData?.party?.data?.partyType === "Witness" ? caseRespondent : respondentDetails,
+        ...(orderFormData?.party?.data?.partyType === "Witness" && { witnessDetails: respondentDetails }),
+        complainantDetails: {
+          name: complainantName,
+          address: complainantAddress,
+        },
+        caseDetails: {
+          caseTitle: caseDetails?.caseTitle,
+          year: new Date(caseDetails).getFullYear(),
+          hearingDate: new Date(orderData?.additionalDetails?.formdata?.dateForHearing || "").getTime(),
+          courtName: courtDetails?.name,
+          courtAddress: courtDetails?.address,
+          courtPhone: courtDetails?.phone,
+          courtId: caseDetails?.courtId,
+          hearingNumber: orderData?.hearingNumber,
+          judgeName: judgeName,
+        },
+        deliveryChannels: {
+          channelName: "",
+          status: "",
+          statusChangeDate: "",
+          fees: 0,
+          feesStatus: "pending",
+        },
+      };
+      break;
+    case "WARRANT":
+      payload = {
+        warrantDetails: {
+          issueDate: orderData?.auditDetails?.lastModifiedTime,
+          caseFilingDate: caseDetails?.filingDate,
+          docType: orderFormValue.warrantType?.code,
+          docSubType: orderFormValue.bailInfo?.isBailable?.code ? "BAILABLE" : "NON_BAILABLE",
+          surety: orderFormValue.bailInfo?.noOfSureties?.code,
+          bailableAmount: orderFormValue.bailInfo?.bailableAmount,
+          templateType: orderFormValue?.warrantSubType?.templateType || "GENERIC",
+          warrantText: orderFormValue?.warrantText?.warrantText || "",
+        },
+        ...(orderFormData?.party?.data?.partyType === "Witness" && { witnessDetails: respondentDetails }),
+        respondentDetails: respondentDetails,
+        caseDetails: {
+          caseTitle: caseDetails?.caseTitle,
+          year: new Date(caseDetails).getFullYear(),
+          hearingDate: new Date(orderData?.additionalDetails?.formdata?.dateOfHearing || "").getTime(),
+          judgeName: judgeName,
+          courtName: courtDetails?.name,
+          courtAddress: courtDetails?.address,
+          courtPhone: courtDetails?.phone,
+          courtId: caseDetails?.courtId,
+        },
+        deliveryChannels: {
+          channelName: "Police",
+          name: "",
+          address: "",
+          phone: "",
+          email: "",
+          status: "",
+          statusChangeDate: "",
+          fees: await getCourtFee(
+            "POLICE",
+            respondentAddress?.[0]?.pincode,
+            orderType === "WARRANT" || orderType === "PROCLAMATION" || orderType === "ATTACHMENT" ? "WARRANT" : orderType,
+            tenantId
+          ),
+          feesStatus: "",
+        },
+      };
+      break;
+    case "PROCLAMATION":
+      payload = {
+        proclamationDetails: {
+          issueDate: orderData?.auditDetails?.lastModifiedTime,
+          caseFilingDate: caseDetails?.filingDate,
+          docSubType: "Proclamation requiring the apperance of a person accused",
+          templateType: "GENERIC",
+          proclamationText: orderFormValue?.proclamationText?.proclamationText || "",
+          partyType: respondentNameData?.partyType?.toLowerCase() || "accused",
+        },
+        ...(orderFormData?.party?.data?.partyType === "Witness" && { witnessDetails: respondentDetails }),
+        respondentDetails: respondentDetails,
+        caseDetails: {
+          caseTitle: caseDetails?.caseTitle,
+          year: new Date(caseDetails).getFullYear(),
+          hearingDate: new Date(orderData?.additionalDetails?.formdata?.dateOfHearing || "").getTime(),
+          judgeName: judgeName,
+          courtName: courtDetails?.name,
+          courtAddress: courtDetails?.address,
+          courtPhone: courtDetails?.phone,
+          courtId: caseDetails?.courtId,
+        },
+        deliveryChannels: {
+          channelName: "Police",
+          name: "",
+          address: "",
+          phone: "",
+          email: "",
+          status: "",
+          statusChangeDate: "",
+          fees: await getCourtFee(
+            "POLICE",
+            respondentAddress?.[0]?.pincode,
+            orderType === "WARRANT" || orderType === "PROCLAMATION" ? "WARRANT" : orderType,
+            tenantId
+          ),
+          feesStatus: "",
+        },
+      };
+      break;
+    case "ATTACHMENT":
+      payload = {
+        attachmentDetails: {
+          issueDate: orderData?.auditDetails?.lastModifiedTime,
+          caseFilingDate: caseDetails?.filingDate,
+          docSubType: "Attachment requiring the apperance of a person accused",
+          templateType: "GENERIC",
+          attachmentText: orderFormValue?.attachmentText?.attachmentText || "",
+          district: orderFormValue?.district?.district || "",
+          village: orderFormValue?.village?.village || "",
+          chargeDays: orderFormValue?.chargeDays?.chargeDays || "",
+          partyType: respondentNameData?.partyType?.toLowerCase() || "accused",
+        },
+        ...(orderFormData?.party?.data?.partyType === "Witness" && { witnessDetails: respondentDetails }),
+        respondentDetails: respondentDetails,
+        caseDetails: {
+          caseTitle: caseDetails?.caseTitle,
+          year: new Date(caseDetails).getFullYear(),
+          hearingDate: new Date(orderData?.additionalDetails?.formdata?.dateOfHearing || "").getTime(),
+          judgeName: judgeName,
+          courtName: courtDetails?.name,
+          courtAddress: courtDetails?.address,
+          courtPhone: courtDetails?.phone,
+          courtId: caseDetails?.courtId,
+        },
+        deliveryChannels: {
+          channelName: "Police",
+          name: "",
+          address: "",
+          phone: "",
+          email: "",
+          status: "",
+          statusChangeDate: "",
+          fees: await getCourtFee(
+            "POLICE",
+            respondentAddress?.[0]?.pincode,
+            orderType === "WARRANT" || orderType === "PROCLAMATION" || orderType === "ATTACHMENT" ? "WARRANT" : orderType,
+            tenantId
+          ),
+          feesStatus: "",
+        },
+      };
+      break;
+    case "BAIL":
+      payload = {
+        respondentDetails: {
+          name: respondentName,
+          address: respondentAddress?.[0],
+          phone: respondentPhoneNo?.[0] || "",
+          email: respondentEmail?.[0] || "",
+          age: "",
+          gender: "",
+        },
+        caseDetails: {
+          title: caseDetails?.caseTitle,
+          year: new Date(caseDetails).getFullYear(),
+          hearingDate: new Date(orderData?.additionalDetails?.formdata?.date || "").getTime(),
+          judgeName: "",
+          courtName: courtDetails?.name,
+          courtAddress: courtDetails?.address,
+          courtPhone: courtDetails?.phone,
+          courtId: caseDetails?.courtId,
+        },
+      };
+      break;
+    case "MISCELLANEOUS_PROCESS":
+      const hearingDate = new Date(orderData?.additionalDetails?.formdata?.dateOfHearing || "").getTime();
+      const taskCaseDetails = {
+        title: caseDetails?.caseTitle,
+        year: new Date(caseDetails).getFullYear(),
+        hearingDate: hearingDate,
+        judgeName: "",
+        courtName: courtDetails?.name,
+        courtAddress: courtDetails?.address,
+        courtPhone: courtDetails?.phone,
+        courtId: caseDetails?.courtId,
+      };
+      const caseNumber = caseDetails?.courtCaseNumber || caseDetails?.cmpNumber || caseDetails?.filingNumber;
+      payload = await _getTaskPayload(taskCaseDetails, orderData, caseDetails?.filingDate, hearingDate, caseNumber, caseDetails?.filingNumber);
+      break;
+    default:
+      break;
+  }
+  if (orderType === "MISCELLANEOUS_PROCESS") return payload;
+  if (Object.keys(payload || {}).length > 0 && !Array.isArray(selectedChannel)) return [payload];
+  else if (Object.keys(payload || {}).length > 0 && Array.isArray(selectedChannel)) {
+    const channelPayloads = await Promise.all(
+      selectedChannel?.map(async (item) => {
+        let clonedPayload = JSON.parse(JSON.stringify(payload));
+
+        const pincode = ["e-Post", "Registered Post", "Via Police"].includes(item?.type)
+          ? item?.value?.pincode
+          : clonedPayload?.respondentDetails?.address?.pincode;
+
+        let courtFees = await getCourtFee(
+          item?.code,
+          pincode,
+          orderType === "WARRANT" || orderType === "PROCLAMATION" || orderType === "ATTACHMENT" ? "WARRANT" : orderType,
+          tenantId
+        );
+
+        if ("deliveryChannels" in clonedPayload) {
+          clonedPayload.deliveryChannels = {
+            ...clonedPayload.deliveryChannels,
+            channelName: channelTypeEnum?.[item?.type]?.type,
+            fees: courtFees,
+            channelCode: channelTypeEnum?.[item?.type]?.code,
+            isPendingCollection: channelTypeEnum?.[item?.type]?.code === "RPAD" ? true : false,
+          };
+
+          let address = {};
+          if (orderType === "WARRANT" || orderType === "PROCLAMATION" || orderType === "ATTACHMENT" || item?.type === "Via Police") {
+            address = {
+              ...item?.value,
+              locality: item?.value?.locality || "",
+              coordinate: {
+                longitude: item?.value?.geoLocationDetails?.longitude,
+                latitude: item?.value?.geoLocationDetails?.latitude,
+              },
+            };
+          } else if (["e-Post", "Registered Post"].includes(item?.type)) {
+            const baseAddress = item?.value || {};
+            address = {
+              ...baseAddress,
+              locality: item?.value?.locality || baseAddress?.locality || "",
+              coordinate: item?.value?.coordinates || baseAddress?.coordinates || {},
+            };
+          } else {
+            const baseAddress = respondentAddress[0] || {};
+            address = {
+              ...baseAddress,
+              coordinate: baseAddress?.coordinates || {},
+            };
+          }
+
+          const phone = item?.type === "SMS" ? item?.value : respondentPhoneNo?.[0] || "";
+          const email = item?.type === "E-mail" ? item?.value : respondentEmail?.[0] || "";
+          const commonDetails = { address, phone, email, age: "", gender: "" };
+
+          clonedPayload.respondentDetails = {
+            ...clonedPayload.respondentDetails,
+            ...commonDetails,
+          };
+
+          if (clonedPayload?.witnessDetails) {
+            clonedPayload.witnessDetails = {
+              ...clonedPayload.witnessDetails,
+              ...commonDetails,
+            };
+          }
+        }
+
+        return clonedPayload;
+      })
+    );
+    return channelPayloads;
+  }
 };
